Improved layout/responsiveness Collaboration Page (#99) + typography(#94)
For the typography system I fixed some issues with components that did
not adopt the new typraphy rules correctly because of legacy styling
issues.

I also improved the images and layout for the collaboration page.
Important improvments are new colours for the illustrations, i have put
the layout inside containers and I made sure that images and text do not
alternate on mobile like they do on desktop.